refactor(diagram): split the SVG recogniser into geometry and text passes
`svg.rs` had grown to 849 lines covering two jobs that share nothing but
a canvas size: reading shapes out of the converted `usvg` tree, and
reading labels out of the source XML because `usvg` dropped them.

    svg/mod.rs        recover(), and the end-to-end tests
    svg/geometry.rs   paths to outlines and connectors, shape naming
    svg/text.rs       the transform chain down to each `<text>` anchor

No behaviour change. Same 40 unit tests, same corpus scores.

Refs #579
